+++ src/SaveObject.js
@@ -55,7 +55,6 @@ function BitburnerSaveObject() {
     this.FconfSettingsSave              = "";
     this.VersionSave                    = "";
     this.AllGangsSave                   = "";
-    this.CorporationResearchTreesSave   = "";
 }
 
 BitburnerSaveObject.prototype.getSaveString = function() {