+++ src/StockMarket/StockMarket.js
@@ -175,6 +175,11 @@ let StockMarket = {}        //Full name to stock object
 let StockSymbols = {}       //Full name to symbol
 let SymbolToStockMap = {};  //Symbol to Stock object
 
+let formatHelpData = {
+    longestName: 0,
+    longestSymbol: 0,
+};
+
 function loadStockMarket(saveString) {
     if (saveString === "") {
         StockMarket = {};
@@ -220,6 +225,11 @@ function initStockSymbols() {
     StockSymbols["Catalyst Ventures"]                       = "CTYS";
     StockSymbols["Microdyne Technologies"]                  = "MDYN";
     StockSymbols["Titan Laboratories"]                      = "TITN";
+
+    for (const key in StockSymbols) {
+        formatHelpData.longestName = key.length > formatHelpData.longestName ? key.length : formatHelpData.longestName;
+        formatHelpData.longestSymbol = StockSymbols[key].length > formatHelpData.longestSymbol ? StockSymbols[key].length : formatHelpData.longestSymbol;
+    }
 }
 
 function initStockMarket() {
@@ -1070,10 +1080,11 @@ function createStockTicker(stock) {
         return;
     }
     var tickerId = "stock-market-ticker-" + stock.symbol;
-    var li = document.createElement("li"), hdr = document.createElement("button");
+    var li = document.createElement("li"), hdr = document.createElement("button"), hdrpre = document.createElement("pre");;
     hdr.classList.add("accordion-header");
     hdr.setAttribute("id", tickerId + "-hdr");
-    hdr.innerHTML = stock.name + "  -  " + stock.symbol + "  -  " + numeralWrapper.format(stock.price, '($0.000a)');
+    hdrpre.textContent = stock.name + "  -  " + stock.symbol + "  -  " + numeralWrapper.format(stock.price, '($0.000a)');
+    hdr.appendChild(hdrpre);
 
     //Div for entire panel
     var stockDiv = document.createElement("div");
@@ -1082,8 +1093,8 @@ function createStockTicker(stock) {
 
     /* Create panel DOM */
     var qtyInput = document.createElement("input"),
-        longShortSelect = document.createElement("select"),
-        orderTypeSelect = document.createElement("select"),
+        longShortSelect = document.createElement("select", {class: "dropdown"}),
+        orderTypeSelect = document.createElement("select", {class: "dropdown"}),
         buyButton = document.createElement("span"),
         sellButton = document.createElement("span"),
         buyMaxButton = document.createElement("span"),
@@ -1102,6 +1113,7 @@ function createStockTicker(stock) {
                           " || (event.keyCode==46) )");
 
     longShortSelect.classList.add("stock-market-input");
+    longShortSelect.classList.add("dropdown");
     longShortSelect.setAttribute("id", tickerId + "-pos-selector");
     var longOpt = document.createElement("option");
     longOpt.text = "Long";
@@ -1113,6 +1125,7 @@ function createStockTicker(stock) {
     }
 
     orderTypeSelect.classList.add("stock-market-input");
+    orderTypeSelect.classList.add("dropdown");
     orderTypeSelect.setAttribute("id", tickerId + "-order-selector");
     var marketOpt = document.createElement("option");
     marketOpt.text = "Market Order";
@@ -1352,19 +1365,15 @@ function updateStockTicker(stock, increase) {
         }
         return;
     }
-    let hdrText = stock.name + " (" + stock.symbol + ") - " + numeralWrapper.format(stock.price, '($0.000a)');
+    const stockPriceFormat = numeralWrapper.format(stock.price, '($0.000a)')
+    let hdrText = `${stock.name}${" ".repeat(1 + formatHelpData.longestName - stock.name.length + (formatHelpData.longestSymbol-stock.symbol.length))}${stock.symbol} -${" ".repeat(10 - stockPriceFormat.length)}${stockPriceFormat}`;
     if (Player.has4SData) {
-        hdrText += " - Volatility: " + numeralWrapper.format(stock.mv, '0,0.00') + "%" +
-                   " - Price Forecast: ";
-        if (stock.b) {
-            hdrText += "+".repeat(Math.floor(stock.otlkMag/10) + 1);
-        } else {
-            hdrText += "-".repeat(Math.floor(stock.otlkMag/10) + 1);
-        }
+        hdrText += ` - Volatility: ${numeralWrapper.format(stock.mv, '0,0.00')}% - Price Forecast: `;
+        hdrText += (stock.b?"+":"-").repeat(Math.floor(stock.otlkMag/10) + 1);
     }
-    hdr.innerText = hdrText;
+    hdr.firstChild.textContent = hdrText;
     if (increase != null) {
-        increase ? hdr.style.color = "#66ff33" : hdr.style.color = "red";
+        hdr.firstChild.style.color = increase ? "#66ff33" : "red";
     }
 }